# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Deal",
    "description" : "Limited time deals",
    "termsOfService" : "http://pokhrel.io/terms/",
    "contact" : {
      "email" : "chintan@pokhrel.io"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://pokhrel.io/deal"
  } ],
  "tags" : [ {
    "name" : "deal",
    "description" : "Limited time deals"
  } ],
  "paths" : {
    "/deal" : {
      "put" : {
        "tags" : [ "deal" ],
        "summary" : "Update an existing deal",
        "operationId" : "updateDeal",
        "requestBody" : {
          "description" : "Deal object, only totalItems and dealExpiry are updated, rest are ignored",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Deal"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid input",
            "content" : { }
          },
          "404" : {
            "description" : "Deal  not found",
            "content" : { }
          }
        },
        "x-codegen-request-body-name" : "body"
      },
      "post" : {
        "tags" : [ "deal" ],
        "summary" : "Add a new deal",
        "operationId" : "addDeal",
        "requestBody" : {
          "description" : "Deal object that needs to be saved",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Deal"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "405" : {
            "description" : "Invalid input",
            "content" : { }
          }
        },
        "x-codegen-request-body-name" : "body"
      }
    },
    "/deal/claim" : {
      "post" : {
        "tags" : [ "deal" ],
        "summary" : "Claim a deal",
        "operationId" : "claimDeal",
        "requestBody" : {
          "description" : "Deal object that needs to be saved",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DealClaim"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "405" : {
            "description" : "Invalid input",
            "content" : { }
          }
        },
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "Register a new user",
        "operationId" : "createUser",
        "requestBody" : {
          "description" : "Created user object",
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "description" : "successful operation",
            "content" : { }
          }
        },
        "x-codegen-request-body-name" : "body"
      }
    }
  },
  "components" : {
    "schemas" : {
      "Deal" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dealEffective" : {
            "type" : "string",
            "description" : "deal effective time in YYYYMMDDHHMMSS"
          },
          "dealExpiry" : {
            "type" : "string",
            "description" : "deal expiry time in YYYYMMDDHHMMSS"
          }
        }
      },
      "DealClaim" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dealId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "userStatus" : {
            "type" : "integer",
            "description" : "User Status",
            "format" : "int32"
          }
        },
        "xml" : {
          "name" : "User"
        }
      }
    }
  }
}